Without weighting  (tested object included)
not sure if we will need weighting honestly. Technically there is actually a weighting system based on order but thats not the most consistent (needs to be tested). Creating a custom weight functionality might take about a day and I have a couple different ideas on it could be implemented Just let me know.